Combine Weather data
The weather data set consists of files for each year. This notebook combines all these files into one. The data set also contains data for stations from all around the world, this is also taken care of by the notebook which selects data of stations that are present in the state of Nevada.